apkb/bot/requirements.txt
# Зависимости Python-бота
python-telegram-bot==20.7
docker==7.0.0
asyncpg==0.29.0
python-dotenv==1.0.0
bleach==6.3.0
